@@ -40,17 +40,15 @@
         "@swc/core": "1.3.56",
         "@swc/jest": "0.2.24",
         "@testing-library/jest-dom": "5.16.5",
-        "@testing-library/react": "10.4.9",
+        "@testing-library/react": "^14.0.0",
         "@types/adm-zip": "^0.5.0",
         "@types/date-fns": "^2.6.0",
-        "@types/enzyme": "3.10.12",
-        "@types/enzyme-adapter-react-16": "1.0.6",
         "@types/ftp": "0.3.33",
         "@types/klaw": "3.0.3",
         "@types/lodash.range": "3.2.7",
         "@types/mousetrap": "1.6.11",
         "@types/react": "17.0.55",
-        "@types/react-dom": "17.0.19",
+        "@types/react-dom": "18.2.6",
         "@types/react-redux": "7.1.25",
         "@types/semver": "7.3.13",
         "@types/shasum": "1.0.0",
@@ -65,11 +63,8 @@
         "date-fns": "2.29.3",
         "electron": "22.3.5",
         "electron-store": "8.1.0",
-        "enzyme": "3.11.0",
-        "enzyme-adapter-react-16": "1.15.7",
-        "enzyme-to-json": "3.6.2",
-        "esbuild": "0.17.18",
-        "esbuild-sass-plugin": "2.9.0",
+        "esbuild": "0.19.2",
+        "esbuild-sass-plugin": "2.13.0",
         "esbuild-style-plugin": "1.6.2",
         "eslint": "8.37.0",
         "eslint-config-airbnb": "19.0.4",
@@ -101,14 +96,14 @@
         "prettier-plugin-tailwindcss": "^0.3.0",
         "prettysize": "2.0.0",
         "protobufjs": "^7.0.0",
-        "react": "16.14.0",
+        "react": "18.2.0",
         "react-bootstrap": "1.6.6",
-        "react-dom": "16.14.0",
+        "react-dom": "18.2.0",
         "react-flip-toolkit": "7.0.17",
-        "react-markdown": "4.3.1",
+        "react-markdown": "8.0.7",
         "react-redux": "7.2.9",
-        "react-resize-detector": "6.7.8",
-        "react-test-renderer": "16.14.0",
+        "react-resize-detector": "9.1.0",
+        "react-test-renderer": "18.2.0",
         "redux": "4.2.1",
         "redux-devtools-extension": "2.13.9",
         "redux-thunk": "2.4.2",
