@@ -17,6 +17,8 @@ module.exports = (disabledMocks = []) => ({
         '^electron$': `${mockDir}/electronMock.ts`,
         '^electron-store$': `${mockDir}/electronStoreMock.ts`,
         '@electron/remote': `${mockDir}/remoteMock.ts`,
+        'react-markdown':
+            '<rootDir>/node_modules/react-markdown/react-markdown.min.js',
         ...(disabledMocks.includes('packageJson')
             ? {}
             : {
@@ -33,7 +35,6 @@ module.exports = (disabledMocks = []) => ({
         'node_modules/(?!(@nordicsemiconductor/pc-nrfconnect-shared)/)',
     ],
     setupFilesAfterEnv: [`${__dirname}/../test/setupTests.ts`],
-    snapshotSerializers: ['enzyme-to-json/serializer'],
     resolver: `${__dirname}/../test/jestResolver.js`,
     modulePathIgnorePatterns: ['dist'],
 });
